fixed: compiler warnings
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@26831 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/VideoRenderers/VideoShaders/ConvolutionKernels.cpp
@@ -206,8 +206,8 @@ void CConvolutionKernel::ToIntFract()
     int integer = value / 255;
     int fract   = value % 255;
 
-    m_intfractpixels[i] = integer;
-    m_intfractpixels[i + m_size * 4] = fract;
+    m_intfractpixels[i] = (uint8_t)integer;
+    m_intfractpixels[i + m_size * 4] = (uint8_t)fract;
   }
 
 #if 0
